timings: display data in local date time
Otherwise we get odd results as GitLab uses UTC. For example,
if time is accounted to a specific day in GitLab UI in CEST time
zone, gitlab sees this as yyyy-mm-ddT22:00:00. This might become
a sunday instead of a monday for example, resulting in wrong
reposts.

Thus, we need to handle times as local times. This is what a
dev usually expects.