cli: fix unfortunate situation with cfg file vs clap
TL;DR: Configurations from config file are treated as arguments
       and parsed by clap.

With the latest addition of config file support, we lost the nice
""`--host` is mandatory" reporting by clap. Now, with this commit,
error reporting is definitely better and less inconsistent than before,
as clap works as before. But now it is not entirely clear to end
users if and how stuff is fetched from the config file, error reporting
is not really existent and yeah... but it should be good enough
and is definitely better than before.