nix/flake: drop unneeded rust-overlay
The Rust overlay is very large, because it carries information about
all possible Rust toolchain versions. This is sometimes useful, but
this tool builds fine with the Rust version in Nixpkgs.

This change reduces the stuff that users have to download quite a bit.
Not only is there now rust-overlay tarball to download, you also use
the default Rust toolchain, which is probably already somewhere on
disk!